+++ Permission/DenyPermission.php
@@ -2,11 +2,9 @@
 
 namespace LoremIpsum\PermissionCheckerBundle\Permission;
 
-use LoremIpsum\PermissionCheckerBundle\AbstractPermission;
-
 class DenyPermission extends AbstractPermission
 {
-    public function isGranted()
+    public function isGranted(): bool
     {
         return false;
     }