+++ Twig/PermissionExtension.php
@@ -2,9 +2,9 @@
 
 namespace LoremIpsum\PermissionCheckerBundle\Twig;
 
-use LoremIpsum\PermissionCheckerBundle\Guardable;
+use LoremIpsum\PermissionCheckerBundle\Model\Guardable;
 use LoremIpsum\PermissionCheckerBundle\Permission\GuardPermission;
-use LoremIpsum\PermissionCheckerBundle\PermissionCheckerInterface;
+use LoremIpsum\PermissionCheckerBundle\Model\PermissionCheckerInterface;
 use Twig\Error\Error;
 
 class PermissionExtension extends \Twig_Extension
@@ -19,18 +19,13 @@ class PermissionExtension extends \Twig_Extension
      */
     protected $actionPermission;
 
-    /**
-     * PermissionExtension constructor.
-     * @param PermissionCheckerInterface $permissionChecker
-     * @param string|null                $actionPermission
-     */
-    public function __construct(PermissionCheckerInterface $permissionChecker, $actionPermission)
+    public function __construct(PermissionCheckerInterface $permissionChecker, ?string $actionPermission)
     {
         $this->permissionChecker = $permissionChecker;
         $this->actionPermission  = $actionPermission;
     }
 
-    public function getFunctions()
+    public function getFunctions(): array
     {
         return [
             new \Twig_Function('hasCreatePermission', [$this, 'hasCreatePermission']),
@@ -42,32 +37,32 @@ class PermissionExtension extends \Twig_Extension
         ];
     }
 
-    public function hasCreatePermission(Guardable $guard)
+    public function hasCreatePermission(Guardable $guard): bool
     {
         return $this->hasGuardPermission($guard, GuardPermission::CREATE);
     }
 
-    public function hasReadPermission(Guardable $guard)
+    public function hasReadPermission(Guardable $guard): bool
     {
         return $this->hasGuardPermission($guard, GuardPermission::READ);
     }
 
-    public function hasUpdatePermission(Guardable $guard)
+    public function hasUpdatePermission(Guardable $guard): bool
     {
         return $this->hasGuardPermission($guard, GuardPermission::UPDATE);
     }
 
-    public function hasDeletePermission(Guardable $guard)
+    public function hasDeletePermission(Guardable $guard): bool
     {
         return $this->hasGuardPermission($guard, GuardPermission::DELETE);
     }
 
-    public function hasGuardPermission(Guardable $guard, $action)
+    public function hasGuardPermission(Guardable $guard, $action): bool
     {
         return $this->permissionChecker->has($guard->getPermission($action));
     }
 
-    public function hasActionPermission($action)
+    public function hasActionPermission($action): bool
     {
         if (! $this->actionPermission) {
             throw new Error("Configure 'lorem_ipsum_permission_checker.default_permission' to be used by hasActionPermission.");